+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/RefenreceServerStart.kt
@@ -1,4 +1,4 @@
-package org.stellar
+package org.stellar.reference
 
 // Used in ServiceRunner
 @JvmOverloads // Annotation required to call from Java with optional argument
