+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/Util.kt
@@ -1,3 +1,3 @@
-package org.stellar
+package org.stellar.reference
 
 class ClientException(message: String) : Exception(message)
