+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/data/Config.kt
@@ -1,4 +1,4 @@
-package org.stellar.data
+package org.stellar.reference.data
 
 import org.stellar.sdk.KeyPair
