+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/jwt/Decoder.kt
@@ -1,10 +1,10 @@
-package org.stellar.jwt
+package org.stellar.reference.jwt
 
 import io.jsonwebtoken.Claims
 import io.jsonwebtoken.JwtParser
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.impl.DefaultJwsHeader
-import org.stellar.data.JwtToken
+import org.stellar.reference.data.JwtToken
 
 object JwtDecoder {
   fun decode(cipher: String, jwtKey: String): JwtToken {
