+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/plugins/TestRoute.kt
@@ -1,4 +1,4 @@
-package org.stellar.plugins
+package org.stellar.reference.plugins
 
 import io.ktor.http.*
 import io.ktor.server.application.*
@@ -8,11 +8,11 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import mu.KotlinLogging
-import org.stellar.ClientException
-import org.stellar.jwt.JwtDecoder
-import org.stellar.sep24.DepositService
-import org.stellar.sep24.Sep24Helper
-import org.stellar.sep24.WithdrawalService
+import org.stellar.reference.ClientException
+import org.stellar.reference.jwt.JwtDecoder
+import org.stellar.reference.sep24.DepositService
+import org.stellar.reference.sep24.Sep24Helper
+import org.stellar.reference.sep24.WithdrawalService
 
 private val log = KotlinLogging.logger {}
