+++ kotlin-reference-server/src/main/kotlin/org/stellar/reference/sep24/DepositService.kt
@@ -1,8 +1,8 @@
-package org.stellar.sep24
+package org.stellar.reference.sep24
 
 import java.math.BigDecimal
 import mu.KotlinLogging
-import org.stellar.data.*
+import org.stellar.reference.data.*
 
 private val log = KotlinLogging.logger {}
